[Silver III] Title: 달나라 토끼를 위한 구매대금 지불 도우미, Time: 112 ms, Memory: 14684 KB -BaekjoonHub

+++ 백준/Silver/17212. 달나라 토끼를 위한 구매대금 지불 도우미/달나라 토끼를 위한 구매대금 지불 도우미.java
@@ -9,15 +9,15 @@ class Main{
 	public static void main(String[] args)throws Exception{
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		int N	= Integer.parseInt(br.readLine());	// 0<=십만
-		long dp[]= new long[N+8];
+		int dp[]= new int[N+8];
 		
 		dp[1] = dp[2] = dp[5] = dp[7] = 1;
 		dp[3] = dp[4] = dp[6] = 2;
 		
 		for(int i=8; i<=N; i++)
 		{
-			long min1 = Math.min(dp[i-1], dp[i-2]);
-			long min2 = Math.min(dp[i-5], dp[i-7]);
+			int min1 = Math.min(dp[i-1], dp[i-2]);
+			int min2 = Math.min(dp[i-5], dp[i-7]);
 			dp[i] = Math.min(min1, min2) + 1;
 		}
 		System.out.print(dp[N]);
